+++ src/app/views/shared/dynamic-library-examples/example-components/highcharts/mock-data.enum.ts
@@ -10,6 +10,7 @@ export enum MockData {
   PrikkedeDataMedToSerier,
   TestData,
   Valgdeltagelse2015,
+  DodsfallHjerteOgKarEtterKommune,
   TestData11,
   TestData12,
   DodsfallHjerteOgKarEtterFylkeFlereAr,