+++ projects/fhi-angular-highcharts/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 > Mar 05, 2026
 
+- :tada: **Enhancement** Added support for Kommune maps. [(#938)](https://github.com/folkehelseinstituttet/Fhi.Frontend.Demo/pull/938)
+
 - :tada: **Enhancement** Add content projection slot for maps to allow custom panels [(#911)](https://github.com/folkehelseinstituttet/Fhi.Frontend.Demo/pull/911)
 
 - :tada: **Enhancement** Add support for diagram custom requirements and disabled warnings [(#913)](https://github.com/folkehelseinstituttet/Fhi.Frontend.Demo/pull/913)